src/triton/utils/requirements.txt
accelerate==0.33.0
aiofiles==24.1.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
altair==5.3.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
async-timeout==4.0.3
attrs==24.2.0
beartype==0.18.5
bitsandbytes==0.41.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
cmake==3.30.2
CoLT5-attention==0.11.1
contourpy==1.3.0
cycler==0.12.1
datasets==3.0.0
decorator==4.4.2
decord==0.6.0
deepspeed==0.14.0
dill==0.3.8
distro==1.9.0
docker-pycreds==0.4.0
einops==0.8.0
einops-exts==0.0.4
einx==0.3.0
et-xmlfile==1.1.0
exceptiongroup==1.2.2
fastapi==0.103.2
ffmpy==0.4.0
filelock==3.15.4
flake8==7.1.1
flash-attn==2.6.3
fonttools==4.53.1
frozendict==2.4.4
frozenlist==1.4.1
fsspec==2024.6.1
gitdb==4.0.11
GitPython==3.1.43
gradio==3.35.2
gradio_client==0.2.9
h11==0.14.0
hjson==3.1.0
httpcore==0.17.3
httpx==0.24.0
huggingface-hub==0.24.6
idna==3.8
imageio==2.35.1
imageio-ffmpeg==0.5.1
iniconfig==2.0.0
Jinja2==3.1.4
joblib==1.4.2
jsonlines==4.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
latex2mathml==3.77.0
libcst==1.1.0
linkify-it-py==2.0.3
lit==18.1.8
local-attention==1.9.15
loguru==0.7.2
markdown-it-py==2.2.0
markdown2==2.5.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mccabe==0.7.0
mdit-py-plugins==0.3.3
mdurl==0.1.2
mgqa==0.0.5
moviepy==1.0.3
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
mypy-extensions==1.0.0
narwhals==1.5.5
navit-paddlepaddle-gpu==0.0.8
networkx==3.3
ninja==1.11.1.1
numpy==1.26.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.6.20
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
openai==1.3.5
opencv-python==4.10.0.84
openpyxl==3.1.5
orjson==3.10.7
packaging==24.1
pandas==2.2.2
peft==0.10.0
pillow==10.4.0
platformdirs==4.2.2
pluggy==1.5.0
portalocker==2.10.1
proglog==0.1.10
protobuf==5.27.4
psutil==6.0.0
py-cpuinfo==9.0.0
pyarrow==17.0.0
pycodestyle==2.12.1
pydantic==1.10.18
pydub==0.25.1
pyflakes==3.2.0
Pygments==2.18.0
pynvml==11.5.3
pyparsing==3.1.4
pytest==8.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.2
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
rich==13.8.0
rpds-py==0.20.0
safetensors==0.4.4
scikit-learn==1.2.2
scipy==1.14.1
semantic-version==2.10.0
sentencepiece==0.1.99
sentry-sdk==2.13.0
setproctitle==1.3.3
shortuuid==1.0.13
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
starlette==0.27.0
sty==1.0.6
svgwrite==1.4.3
sympy==1.13.2
tabulate==0.9.0
tensorboardX==2.6.2.2
threadpoolctl==3.5.0
tiktoken==0.7.0
timeout-decorator==0.5.0
timm==0.9.16
tokenizers==0.19.1
tomli==2.0.1
toolz==0.12.1
paddlepaddle-gpu==2.1.2
paddlepaddle-gpuaudio==2.1.2
TorchFix==0.6.0
paddlepaddle-gpuvision==0.16.2
tqdm==4.66.5
transformers==4.44.2
triton==2.1.0
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2024.1
uc-micro-py==1.0.3
urllib3==2.2.2
uvicorn==0.30.6
validators==0.34.0
vector-quantize-pypaddlepaddle-gpu==1.17.1
wandb==0.17.7
wavedrom==2.0.3.post3
websockets==13.0
XlsxWriter==3.2.0
xxhash==3.5.0
yarl==1.9.4
zetascale==2.7.3
